server2/requirements.txt
Flask==0.12
watson_developer_cloud==0.23.0
python-dotenv==0.6.2
Flask-SocketIO==2.8.2
